// src/app/application/my-modules/inspection/ipqa/shared/model/ipqa.ts
export class IpqaModel {
    checkDate: string;
    checkPerson: string;
    banbie: string;
    address: string;
    checklist_cn: string;
    exceptionDesc: string;
    pictures: string;
    handler?: string;
    actionDesc?: string;
    actionStatus?: string;
    actionPictures?: string;
    actionDate?: string;
}